+++ CHANGELOG.md
@@ -9,3 +9,4 @@
 - [2026-05-16 22:42] data/data/com.termux/files/home/godmode/lib/detectors.py
 - [2026-05-16 22:43] data/data/com.termux/files/home/godmode/test/canary_probe.py
 - [2026-05-16 22:46] data/data/com.termux/files/home/godmode/main.py
+- [2026-05-16 23:43] home/godmode/engine/racer.py